refactor(execution_error): move it to ast so it can be used in dto
